Use insecure 3sec for all tests, skip paper key on signup.

Local speed ups:

teams:  126s => 73s
chat: 186s => 112s

(engine already had both of these optimizations)