@maxtaco 2017-09-13
Excellent!!!!

On Wed, Sep 13, 2017 at 5:07 PM Patrick Crosby <notifications@github.com>
wrote:

> Merged #8438 <https://github.com/keybase/client/pull/8438>.
>
> —
> You are receiving this because your review was requested.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/keybase/client/pull/8438#event-1248082777>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AA05_0n8wjD8UY1YEQbanT9GjltrlQBzks5siEQSgaJpZM4PWqMs>
> .
>