@maxtaco 2017-09-13
We might need to back this out due to the vendoring problems it introduces.  We should not be requiring libkb to vendor in `triplesec-insecure`.

cc: @jzila