@maxtaco 2017-09-14
It's such a dangerous library, I really don't want it spreading virally through the inclusion graph. I made a conscious effort to keep it local to engine when I introduced it

I think there are other ways to achieve the same result.  Just copy/pasta test code in the 3 directories where it was introduced.  I find that much preferable to introducing a new dependency that needs to be vendored.