@jzila 2017-09-14
Ideally these `test_common` libraries would exist in a separate package that is imported in `_test.go` files only, and then you can do all of this without affecting the inclusion graph.