@patrickxb 2017-09-14
That's what `kbtest` was intended for ages ago.

I'll change test_common.go to not import it.