@maxtaco 2017-09-14
Thanks! I realize it's paranoia, but this would be a bad bug if this code got out and into prod.