@jzila 2017-09-14
Long term would it be very difficult to move `libkb/test_common.go` and `externals/test_common.go` into `kbtest`?